refactor(christmas): 로직 변경
