+++ src/main/java/christmas/constant/StarDiscount.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public enum StarDiscount {
     SPECIAL_DISCOUNT("특별 할인", 1000, Arrays.asList(3, 10, 17, 24, 25, 31)),
-    EMPTY("없음", 0, Collections.EMPTY_LIST);
+    EMPTY("없음", 0, Collections.emptyList());
 
     private final String discountCategory;
     private final int discountPrice;
