+++ src/main/java/christmas/domain/event/BadgeEvent.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.event;
 
 import christmas.constant.Badge;
