+++ src/main/java/christmas/domain/event/GiftEvent.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.event;
 
 import christmas.constant.Menu;
 import christmas.constant.NumberConstant;
